requirements.txt
anndata==0.9.2
click==8.1.7
einops==0.6.1
fairseq==0.12.2
hdf5plugin==4.1.3  # needed by older version of anndata
louvain==0.8.0
numba==0.57.1
numpy==1.24.4
pandas==1.5.3
pytorch-lightning==2.0.4
scikit-learn==1.2.2
scipy==1.11.1
tqdm==4.65.0
wandb==0.15.5
adjustText==0.8
### GEARS
dcor==0.6
networkx==2.8.4
### Single cell related
rapids_singlecell==0.7.2
scanpy==1.9.3
scib==1.1.3
scprep==1.2.3
scvi-tools==1.0.0
### Models
transformers==4.30.2
### Tests
pytest-subtests==0.11.0
pytest==7.4.0
